requirements.txt
wheel
boto3
orjson
brotli
bs4
celery
cryptography
elasticsearch==7.13.0
django==4.0.6
djangorestframework
django-cors-headers
django-crispy-forms
django-widget-tweaks
django-storages
django-ckeditor
django-redis
djangorestframework-simplejwt
drf-recaptcha
facebook-sdk
geopandas
gspread
google-auth
httplib2
huey
lightgbm
# mysqlclient
numpy
oauth2client
pandas
psycopg2-binary
pycryptodome
pyinstrument
pyppeteer
python-dateutil
redis
requests
stripe
sentry-sdk
shapely
SQLAlchemy
tzdata
scikit-learn
